Revert "feature(cassandra-stress): move to run on docker base image"
This reverts commit d580a9900a06dc0a37a96e0bbb940d40442af810.
