+++ sdcm/utils/docker_remote.py
@@ -110,9 +110,3 @@ class RemoteDocker(BaseNode):
         prefix = "sudo" if node.is_docker else ""
         node.remoter.run(
             f'{prefix} docker pull {image}', verbose=True)
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        self.kill()
