+++ sdcm/ycsb_thread.py
@@ -26,7 +26,7 @@ from sdcm.utils import alternator
 from sdcm.utils.common import FileFollowerThread
 from sdcm.utils.docker_remote import RemoteDocker
 from sdcm.utils.common import generate_random_string
-from sdcm.stress.base import format_stress_cmd_error, DockerBasedStressThread
+from sdcm.stress_thread import format_stress_cmd_error, DockerBasedStressThread
 
 LOGGER = logging.getLogger(__name__)
